+implement local storing, add, edit, delete function

+++ app/components/Dashboard/risk_assesment/save_family_risk_profile.js
@@ -10,6 +10,8 @@ export default class SaveFamilyRiskProfile extends Component {
 
         this.state = {
             family_profile_id: 0,
+            local_storage_id: 0,
+            sync_status: 0,
             members_count: "",
             vulnerable_members_count: "",
             vulnerability_nature: ""
@@ -23,6 +25,8 @@ export default class SaveFamilyRiskProfile extends Component {
         if (data != "none") {
             this.setState({
                 family_profile_id: data.family_profile_id.toString(),
+                local_storage_id: data.local_storage_id,
+                sync_status: data.sync_status,
                 members_count: data.members_count.toString(),
                 vulnerable_members_count: data.vulnerable_members_count.toString(),
                 vulnerability_nature: data.vulnerability_nature
@@ -30,6 +34,8 @@ export default class SaveFamilyRiskProfile extends Component {
         } else {
             this.setState({
                 family_profile_id: 0,
+                local_storage_id: 0,
+                sync_status: 0,
                 members_count: "",
                 vulnerable_members_count: "",
                 vulnerability_nature: ""
@@ -39,6 +45,8 @@ export default class SaveFamilyRiskProfile extends Component {
 
     saveFamilyRiskProfile() {
         const { family_profile_id,
+            local_storage_id,
+            sync_status,
             members_count,
             vulnerable_members_count,
             vulnerability_nature } = this.state
@@ -51,6 +59,8 @@ export default class SaveFamilyRiskProfile extends Component {
             },
             body: JSON.stringify({
                 family_profile_id: family_profile_id,
+                local_storage_id: local_storage_id,
+                sync_status: sync_status,
                 members_count: members_count,
                 vulnerable_members_count: vulnerable_members_count,
                 vulnerability_nature: vulnerability_nature
@@ -59,8 +69,42 @@ export default class SaveFamilyRiskProfile extends Component {
             .then((responseJson) => {
                 console.log(responseJson)
                 if (responseJson.status == true) {
-                    this.props.navigation.navigate('modify_family_risk');
                     ToastAndroid.show(responseJson.message, ToastAndroid.SHORT);
+                    let data_container = Storage.getItem('RiskAssessmentFamilyRiskProfile')
+                    let updated_data = []
+                    data = {
+                        family_profile_id: family_profile_id,
+                        local_storage_id: 1,
+                        sync_status: 3,
+                        members_count: members_count,
+                        vulnerable_members_count: vulnerable_members_count,
+                        vulnerability_nature: vulnerability_nature
+                    }
+                    data_container.then(response => {
+                        if (response == null) {
+                            Storage.removeItem("RiskAssessmentFamilyRiskProfile")
+                            Storage.setItem("RiskAssessmentFamilyRiskProfile", [data])
+                        } else {
+                            let temp = response
+                            temp.push(data)
+                            let counter = 0
+                            temp.forEach((value) => {
+                                counter += 1
+                                updated_data.push({
+                                    family_profile_id: value.family_profile_id,
+                                    local_storage_id: counter,
+                                    sync_status: 3,
+                                    members_count: value.members_count,
+                                    vulnerable_members_count: value.vulnerable_members_count,
+                                    vulnerability_nature: value.vulnerability_nature
+                                })
+                            });
+                            Storage.removeItem("RiskAssessmentFamilyRiskProfile")
+                            Storage.setItem("RiskAssessmentFamilyRiskProfile", updated_data)
+                        }
+                    });
+
+                    this.props.navigation.navigate('modify_family_risk');
                 } else {
                     ToastAndroid.show(responseJson.message, ToastAndroid.SHORT);
                 }
@@ -68,21 +112,67 @@ export default class SaveFamilyRiskProfile extends Component {
             .catch((error) => {
                 data = {
                     family_profile_id: family_profile_id,
+                    local_storage_id: local_storage_id,
+                    sync_status: 1,
                     members_count: members_count,
                     vulnerable_members_count: vulnerable_members_count,
                     vulnerability_nature: vulnerability_nature
                 }
                 let offline_data = Storage.getItem("RiskAssessmentFamilyRiskProfile");
                 offline_data.then(response => {
-                    if (response == null) {
-                        Storage.removeItem("RiskAssessmentFamilyRiskProfile")
-                        Storage.setItem("RiskAssessmentFamilyRiskProfile", [data])
+                    if (local_storage_id == 0) {
+                        if (response == null) {
+                            Storage.removeItem("RiskAssessmentFamilyRiskProfile")
+                            Storage.setItem("RiskAssessmentFamilyRiskProfile", [data])
+                        } else {
+                            let temp = response
+                            temp.push(data)
+                            let updated_data = []
+                            let counter = 0
+                            temp.forEach((value) => {
+                                counter += 1
+                                updated_data.push({
+                                    family_profile_id: value.family_profile_id,
+                                    local_storage_id: counter,
+                                    sync_status: value.sync_status,
+                                    members_count: value.members_count,
+                                    vulnerable_members_count: value.vulnerable_members_count,
+                                    vulnerability_nature: value.vulnerability_nature
+                                })
+                            });
+                            Storage.removeItem("RiskAssessmentFamilyRiskProfile")
+                            Storage.setItem("RiskAssessmentFamilyRiskProfile", updated_data)
+                        }
                     } else {
                         let temp = response
-                        temp.push(data)
+                        let updated_data = []
+                        let counter = 0
+                        temp.forEach((value) => {
+                            counter += 1
+                            if (local_storage_id == value.local_storage_id) {
+                                updated_data.push({
+                                    family_profile_id: family_profile_id,
+                                    local_storage_id: counter,
+                                    sync_status: 2,
+                                    members_count: members_count,
+                                    vulnerable_members_count: vulnerable_members_count,
+                                    vulnerability_nature: vulnerability_nature
+                                })
+                            } else {
+                                updated_data.push({
+                                    family_profile_id: value.family_profile_id,
+                                    local_storage_id: counter,
+                                    sync_status: value.sync_status,
+                                    members_count: value.members_count,
+                                    vulnerable_members_count: value.vulnerable_members_count,
+                                    vulnerability_nature: value.vulnerability_nature
+                                })
+                            }
+                        });
                         Storage.removeItem("RiskAssessmentFamilyRiskProfile")
-                        Storage.setItem("RiskAssessmentFamilyRiskProfile", temp)
+                        Storage.setItem("RiskAssessmentFamilyRiskProfile", updated_data)
                     }
+
                 })
                 this.props.navigation.navigate('modify_family_risk');
             });
